Moved components to dedicated folders and moved tests to their own folder.
