+++ client/src/components/pet-components/CheckInToggle.jsx
