+++ client/src/components/pet-components/CheckedInPetsList.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { Navbar } from "./Navbar";
+import { Navbar } from "../Navbar";
 import { PetListView } from "./PetListView";
 
 export function CheckedInPetsList({ baseURL }) {
