+++ client/src/components/pet-components/PetProfileIconBar.jsx
