+++ client/src/main.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import { BrowserRouter } from "react-router-dom";
 import { createRoot } from "react-dom/client";
-import App from "./App.jsx";
+import { App } from "./App.jsx";
 
 createRoot(document.getElementById("root")).render(
   <BrowserRouter>
