+++ client/tests/AllPetsList.test.jsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import { AllPetsList } from "./AllPetsList";
+import { AllPetsList } from "../src/components/pet-components/AllPetsList";
 const baseURL = import.meta.env.VITE_API_URL;
 
 describe("All Pets List", () => {
