+++ client/tests/App.test.jsx
@@ -1,16 +1,14 @@
 import { render, screen } from "@testing-library/react";
 
-import App from "./App";
+import { App } from "../src/App";
 import { BrowserRouter } from "react-router-dom";
 
-describe.skip("App", () => {
+describe("App", () => {
   it("renders the main app", () => {
     render(
       <BrowserRouter>
         <App />
       </BrowserRouter>
     );
-
-    screen.debug();
   });
 });
