+++ client/tests/EditPetForm.test.jsx
@@ -1,7 +1,7 @@
 // to do: figure out how to test for default information rendering without props
 
 import { render } from "@testing-library/react";
-import { EditPetForm } from "./EditPetForm";
+import { EditPetForm } from "../src/components/pet-components/EditPetForm";
 const baseURL = import.meta.env.VITE_API_URL;
 
 describe("Edit Pet Form", () => {
