+++ client/tests/Homepage.test.jsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import { Homepage } from "./Homepage";
+import { Homepage } from "../src/components/Homepage";
 
 describe("Homepage", () => {
   const page = render(<Homepage />);
