+++ client/tests/Navbar.test.jsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import { Navbar } from "./Navbar";
+import { Navbar } from "../src/components/Navbar";
 
 describe("Navbar", () => {
   const navbar = render(<Navbar />);
