+++ client/tests/PetProfile.test.jsx
@@ -1,8 +1,8 @@
 import { render } from "@testing-library/react";
-import { PetProfile } from "./PetProfile";
+import { PetProfile } from "../src/components/pet-components/PetProfile";
 const baseURL = import.meta.env.VITE_API_URL;
 
-describe.skip("Pet Profile", () => {
+describe("Pet Profile", () => {
   const profile = render(<PetProfile baseURL={baseURL} />);
   it("renders the profile", () => {
     expect(profile);
