+++ client/tests/PetProfileTabs.test.jsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import { PetProfileTabs } from "./PetProfileTabs";
+import { PetProfileTabs } from "../src/components/pet-components/PetProfileTabs";
 
 describe("Pet Profile Tabs", () => {
   const tabs = render(<PetProfileTabs />);
